requirements/production.txt
-r base.txt
sentry-sdk[django]==1.38.0
django-silk==5.0.4
django-ratelimit==4.1.0
django-redis==5.4.0
